Reorganize `users` app services into a package
