+++ apps/articles/filters.py
@@ -10,7 +10,7 @@ from django_filters.filters import (
 from django_filters.widgets import DateRangeWidget
 from django_select2.forms import Select2TagWidget
 
-from users.services import get_all_users
+from users.services.services import get_all_users
 
 from .models import Article
 from .services import (
