+++ apps/users/signals.py
@@ -1,4 +1,4 @@
-from .services import create_user_profile
+from .services.services import create_user_profile
 
 
 def create_profile(sender, instance, created, **kwargs):
